Project/LED_Matrix_Display/requirement.txt
RPi.GPIO == 0.7.1
rpi_ws281x == 5.0.0
adafruit-circuitpython-neopixel == 6.3.11
Pillow == 10.1.0